2. Dashboard/requirements.txt
dash>=2.18
dash-bootstrap-components>=1.6
plotly>=5.24
apscheduler>=3.10
python-dotenv>=1.0
requests>=2.31
pandas>=2.2
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
flask>=3.0
